+++ src/Data/Promotion/Prelude/Identity.hs
@@ -1,0 +1,20 @@
+-----------------------------------------------------------------------------
+-- |
+-- Module      :  Data.Promotion.Prelude.Identity
+-- Copyright   :  (C) 2018 Ryan Scott
+-- License     :  BSD-style (see LICENSE)
+-- Maintainer  :  Richard Eisenberg (rae@cs.brynmawr.edu)
+-- Stability   :  experimental
+-- Portability :  non-portable
+--
+-- Exports the defunctionalization symbols for the 'Identity' data type.
+--
+-----------------------------------------------------------------------------
+
+module Data.Promotion.Prelude.Identity (
+  -- * Defunctionalization symbols
+  IdentitySym0, IdentitySym1,
+  RunIdentitySym0, RunIdentitySym1
+  ) where
+
+import Data.Singletons.Prelude.Identity