+++ src/Data/Singletons/Prelude/Show.hs
@@ -57,7 +57,7 @@ import           Data.Proxy
 import           Data.Singletons.Internal
 import           Data.Singletons.Prelude.Base
 import           Data.Singletons.Prelude.Instances
-import           Data.Singletons.Prelude.List
+import           Data.Singletons.Prelude.List.Internal
 import           Data.Singletons.Prelude.Ord
 import           Data.Singletons.Prelude.Semigroup.Internal
 import           Data.Singletons.Promote